backend/requirements.txt
# BuffetGPT Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Vision (optional - use fallback if not installed)
Pillow>=10.0.0

# Google Gemini Vision (free tier - optional)
google-genai

# YOLOv8 (optional - for local detection)
# ultralytics>=8.0.0
